Player faces updated + Vignette
When the player is female, the correct image is now shown.
Added vignette overlay.
Added sand explosion particle effects to the loaded button and play button.
Before finishing the particle system I created a function to check for valid objects. Its not really needed at the minute but will be useful in the future, so it stays.
Ive split the play button into parts as it will have an animated background.