UI Update (Board, Sand, Player Faces, Default Font). Console commands now only in debug mode
UI Update: New board with a shadow and some cut edges to give the illusion of the board sitting in the sand. (NOTE: Im also going to have to have the wE part of the board duplicated. At the minute the counter is sent behind the board. However as the stack increases, the counters go behind the board, occluding them partly).
Tightened up the popup message.
Dice fade in and out now.
Fixed start up face in case player 1 is female.