Lightning Update + Play Again + Score blocks + Pointer + Jewels
Lightning is now more "realistic". After it hits a random amount of gems are scattered from the lightning endpoints.
TODO: fix the offset (currently the lightning is static with regards to position)
Added play again system. Score blocks are updated on wins, showing current score.
Added custom pointer images and particle effect. When player 2 is CPU is pointer is hidden, but the particle isnt. This was on purpose due to a bug when setting the alpha of particles to 0 and back will cause the particle system to change (ignores things like the brightness of the particle as well as the speed it fades - Very strange).
Moved the dice towards the right of the screen to allow space for the score blocks.