Intro screen decopy
The intial image was taken from some website, so..
Created a new intro with my own images and particle effects.
Resized some of the (intro) UI based on new look.

Changed the score blocks.

Ive added the new board image, split for making the counters fit between the ground and the board. Still to be implemented.

Updated the rain loop to be a little quieter.